add 64bit libs and do according adjustments
	modified:   linden/indra/cmake/LLWindow.cmake
	modified:   linden/indra/develop.py
	modified:   linden/indra/newview/linux_tools/wrapper.sh
	modified:   linden/indra/newview/llappviewerlinux.cpp
	modified:   linden/indra/newview/viewer_manifest.py
	modified:   linden/install.xml

+++ linden/indra/newview/llappviewerlinux.cpp
@@ -256,7 +256,7 @@ static inline BOOL do_elfio_glibc_backtrace()
 	size_t btpos;
 	for (btpos = 0; btpos < btsize; ++btpos)
 	{
-		fprintf(StraceFile, "%d:\t", btpos);
+		fprintf(StraceFile, "%ld:\t", (long)btpos);
 		int symidx;
 		for (symidx = 0; symidx < nSymNo; ++symidx)
 		{
